merge (dev): add dev to main

+++ server/src/command/graphic/mct.c
@@ -9,8 +9,8 @@
 
 int mct(client_t *client, server_t *server, UNUSED char **args)
 {
-    for (uint x = 0; x < server->map->width; x++) {
-        for (uint y = 0; y < server->map->height; y++) {
+    for (uint x = 0; x < (uint)server->map->width; x++) {
+        for (uint y = 0; y < (uint)server->map->height; y++) {
             print_bct(client, map_get_tile(server->map, x, y), x, y);
         }
     }
